+++ src/deep-equal.js
@@ -138,7 +138,11 @@ function objEquiv(a, b, opts) {
   return typeof a === typeof b;
 }
 
-module.exports = deepEqual;
+module.exports = {
+  deepEqual,
+  // Maximum collection length to use deepEqual()
+  maxCollSizeForDeepEqual: 6
+};
 
 // The license for node-deep-equal, on which the above code is based, is as
 // follows: